refactor(ui): split landing categories client shell
